Add files via upload
The bank management system is an application for maintaining a person’s account in a bank . 
This project is designed to simplify the job of the bank employer considerably simpler when 
preserving the records of the client. It allows the consumer to make simple deposits or 
withdrawals and makes the employer's process more user-friendly. It can offer the finest 
security this way. The process of depositing and withdrawing funds can be carried out by the 
employer in the safest possible manner. This application is safe and uses a three-step 
authentication method, including a login pin, an OTP sent through email, and Radio Frequency 
Identification (RFID), to make sure that no unauthorised users may alter the user information 
and that only legitimate bank employees can access it.